pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibe-scanner"
version = "0.1.0"
description = "AI-generated code detection tool and Python library"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Vibe Scanner Contributors"}
]
keywords = ["ai", "code-analysis", "detection", "static-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "tree-sitter>=0.20.0,<0.21.0",
    "gitpython>=3.1.0,<4.0.0",
    "click>=8.0.0,<9.0.0",
    "rich>=13.0.0,<14.0.0",
    "pyyaml>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
stats = [
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0,<2.0.0",
]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "hypothesis>=6.0.0,<7.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "black>=23.0.0,<24.0.0",
    "mypy>=1.0.0,<2.0.0",
    "ruff>=0.1.0,<0.2.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
vibe-scanner = "vibe_scanner.cli:main"

[project.urls]
Homepage = "https://github.com/vibe-scanner/vibe-scanner"
Repository = "https://github.com/vibe-scanner/vibe-scanner"
Issues = "https://github.com/vibe-scanner/vibe-scanner/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "tree_sitter.*",
    "git.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=vibe_scanner",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src/vibe_scanner"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
